Ads-sim-rag/requirements.txt
praw
openai
chromadb
python-dotenv
tqdm